refactor: kebab case exo filenames
