+++ packages/orchestration/src/service.js
@@ -3,8 +3,8 @@
 import { V as E } from '@agoric/vow/vat.js';
 import { M } from '@endo/patterns';
 import { Shape as NetworkShape } from '@agoric/network';
-import { prepareChainAccountKit } from './exos/chainAccountKit.js';
-import { prepareICQConnectionKit } from './exos/icqConnectionKit.js';
+import { prepareChainAccountKit } from './exos/chain-account-kit.js';
+import { prepareICQConnectionKit } from './exos/icq-connection-kit.js';
 import {
   makeICAChannelAddress,
   makeICQChannelAddress,
