+++ packages/orchestration/src/types.ts
@@ -3,8 +3,8 @@
 export type * from './chain-info.js';
 export type * from './cosmos-api.js';
 export type * from './ethereum-api.js';
-export type * from './exos/chainAccountKit.js';
-export type * from './exos/icqConnectionKit.js';
+export type * from './exos/chain-account-kit.js';
+export type * from './exos/icq-connection-kit.js';
 export type * from './orchestration-api.js';
 export type * from './service.js';
 export type * from './vat-orchestration.js';
